requirements.txt
#pip install torch==1.10.1+cu113 torchvision==0.11.2+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html
pillow
gdown
fastapi
Jinja2
uvicorn
python-multipart